Remove package data.
These declarations aren't correct. `package_data` is meant to declare what data to include in what Python packages. So these declarations say, for example, to include "data/*" in the "data" Python package. These declarations are meant for the installation, so they're essentially saying, create a "data" package and copy all of these files to "site-packages/data/*". That's not appropriate, as this package doesn't own that Python namespace.